[reservation.scheduling.timing] Moved "TimeStampUTC".
